+++ src/ExAs/Utils/SystemExtensions/SystemFunctions.cs
@@ -70,17 +70,5 @@ namespace ExAs.Utils.SystemExtensions
                 return false;
             }
         }
-
-        public static int Compare_NullAware<T>(this T first, T other)
-            where T : IComparable<T>
-        {
-            if (first == null)
-            {
-                if (other == null)
-                    return 0;
-                return other.CompareTo(first);
-            }
-            return first.CompareTo(other);
-        }
     }
 }