+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 ## New in 0.4.0
 * added datetime assertions: is-after (#35), is-before (#36)
+* added list is-equal assertion (#44)
 
 ## New in 0.3.0
 * renamed the is-bigger-than assertion to is-greater-than (#26)